refactor(broker): split submitter.rs into directory layout (C36 of PR-3)
Pure mechanical move. submitter.rs (1039 lines) becomes:

- submitter/mod.rs (~27 lines): module decls + pub use service::Submitter.
- submitter/error.rs (~55 lines): SubmitterErr enum + coded_error_impl!.
- submitter/service.rs (~1002 lines): Submitter<P> struct + impl,
  BrokerService impl, and the original test module.

External path crate::submitter::Submitter preserved via the re-export.
SubmitterErr stays pub but has no external consumers, so it is not
re-exported from mod.rs.

Skipped a separate types.rs per the canonical plan's note — nothing in
this service is worth extracting on its own.
